saving pics implemented
git-svn-id: https://localhost/igvc@274 b8439097-6f30-0410-869d-b57f63bf1bb9

+++ 2008/candi/vision/vision_barrels.cc
@@ -332,7 +332,7 @@ void visFindBarrels(void) {
 					//bigWidth*(1 - 2*barrelEdgeNoiseTolerance) == smallWidth
 					
 					int smallWidth = rightBarrelEdgeX - leftBarrelEdgeX + 1;
-					int bigWidth = smallWidth/(1 - 2*barrelEdgeNoiseTolerance);
+					int bigWidth = (int)(smallWidth/(1 - 2*barrelEdgeNoiseTolerance));
 					int lengthToGrowSides = (bigWidth-smallWidth)/2;
 					//printf("DEBUG: lengthToGrowSides=%d\n", lengthToGrowSides);
